+++ waspc/src/Wasp/Generator/NpmInstall.hs
@@ -22,6 +22,7 @@ import qualified Wasp.Job.Kind as Kind
 import qualified Wasp.Job.Node as Node
 import qualified Wasp.Job.Output as Job.Output
 import qualified Wasp.Job.Output.Event as Event
+import Wasp.Process (InputMode (NoInput))
 import Wasp.Project.Common (WaspProjectDir, nodeModulesDirInWaspProjectDir)
 import Wasp.Util (secondsToMicroSeconds)
 import qualified Wasp.Util.IO as IOUtil
@@ -77,7 +78,7 @@ installNpmDependenciesAndReport :: Path' Abs (Dir WaspProjectDir) -> Job.Job ()
 installNpmDependenciesAndReport projectDir = do
   Job.emitJobOutput Event.Stdout "Starting npm install\n"
   Job.withBackgroundOutputWorker reportInstallationProgress $
-    Node.runChecked [] projectDir "npm" ["install"]
+    Node.runChecked NoInput [] projectDir "npm" ["install"]
 
 reportInstallationProgress :: (Event.JobOutputKind -> T.Text -> IO ()) -> IO ()
 reportInstallationProgress emit =